+++ lib/rspec/core/example.rb
@@ -12,9 +12,7 @@ module RSpec
       #
       # Used to define methods that delegate to this example's metadata
       def self.delegate_to_metadata(*keys)
-        keys.each do |key|
-          define_method(key) {@metadata[key]}
-        end
+        keys.each { |key| define_method(key) { @metadata[key] } }
       end
 
       delegate_to_metadata :full_description, :execution_result, :file_path, :pending, :location
@@ -83,7 +81,7 @@ module RSpec
 
         begin
           unless pending
-            with_around_hooks do
+            with_around_each_hooks do
               begin
                 run_before_each
                 @example_group_instance.instance_eval(&@example_block)
@@ -119,7 +117,7 @@ module RSpec
       # Wraps the example block in a Proc so it can invoked using `run` or
       # `call` in [around](../Hooks#around-instance_method) hooks.
       def self.procsy(metadata, &proc)
-        Proc.new(&proc).extend(Procsy).with(metadata)
+        proc.extend(Procsy).with(metadata)
       end
 
       # @private
@@ -153,8 +151,8 @@ module RSpec
       end
 
       # @private
-      def around_hooks
-        @around_hooks ||= example_group.around_hooks_for(self)
+      def around_each_hooks
+        @around_each_hooks ||= example_group.around_each_hooks_for(self)
       end
 
       # @private
@@ -175,13 +173,28 @@ module RSpec
         finish(reporter)
       end
 
+      # @private
+      def instance_eval(&block)
+        @example_group_instance.instance_eval(&block)
+      end
+
+      # @private
+      def instance_eval_with_rescue(&block)
+        @example_group_instance.instance_eval_with_rescue(&block)
+      end
+
+      # @private
+      def instance_eval_with_args(*args, &block)
+        @example_group_instance.instance_eval_with_args(*args, &block)
+      end
+
     private
 
-      def with_around_hooks(&block)
-        if around_hooks.empty?
+      def with_around_each_hooks(&block)
+        if around_each_hooks.empty?
           yield
         else
-          @example_group_class.run_around_each_hooks(self, Example.procsy(metadata, &block)).call
+          @example_group_class.run_around_each_hooks(self, Example.procsy(metadata, &block))
         end
       end
 
@@ -234,10 +247,9 @@ module RSpec
       end
 
       def assign_auto_description
+        return unless RSpec.configuration.expecting_with_rspec?
         if metadata[:description].empty? and !pending?
-          if RSpec.configuration.expecting_with_rspec?
-            metadata[:description] = RSpec::Matchers.generated_description
-          end
+          metadata[:description] = RSpec::Matchers.generated_description
         end
         RSpec::Matchers.clear_generated_description
       end
@@ -245,7 +257,6 @@ module RSpec
       def record(results={})
         execution_result.update(results)
       end
-
     end
   end
 end