+++ lib/rspec/core/formatters/base_text_formatter.rb
@@ -52,8 +52,17 @@ module RSpec
         end
 
         def dump_profile
-          sorted_examples = examples.sort_by { |example| example.execution_result[:run_time] }.reverse.first(10)
-          output.puts "\nTop #{sorted_examples.size} slowest examples:\n"
+          sorted_examples = examples.sort_by {|example|
+            example.execution_result[:run_time] }.reverse.first(10)
+
+          total, slows = [examples, sorted_examples].map {|exs|
+            exs.inject(0.0) {|i, e| i + e.execution_result[:run_time] }}
+
+          time_taken = slows / total
+          percentage = '%.1f' % ((time_taken.nan? ? 0.0 : time_taken) * 100)
+
+          output.puts "\nTop #{sorted_examples.size} slowest examples (#{format_seconds(slows)} seconds, #{percentage}% of total time):\n"
+
           sorted_examples.each do |example|
             output.puts "  #{example.full_description}"
             output.puts cyan("    #{red(format_seconds(example.execution_result[:run_time]))} #{red("seconds")} #{format_caller(example.location)}")