+++ spec/rspec/core/configuration_options_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 require 'ostruct'
 
-describe RSpec::Core::ConfigurationOptions do
+describe RSpec::Core::ConfigurationOptions, :fakefs do
   include ConfigOptionsHelper
 
   it "warns when HOME env var is not set", :unless => (RUBY_PLATFORM == 'java') do
@@ -350,6 +350,11 @@ describe RSpec::Core::ConfigurationOptions do
         options = parse_options("-O", "./custom.opts")
         options[:format].should be_nil
         options[:color].should be_true
+      end
+      it "parses -e 'full spec description'" do
+        File.open("./custom.opts", "w") {|f| f << "-e 'The quick brown fox jumps over the lazy dog'"}
+        options = parse_options("-O", "./custom.opts")
+        options[:full_description].should == /The\ quick\ brown\ fox\ jumps\ over\ the\ lazy\ dog/
       end
     end
   end